refactor: store를 feature에 이관
