+++ src/entity/comment/model/selectedComment.ts
