+++ src/entity/post/model/selectedPost.ts
