+++ src/entity/user/model/selectedUser.ts
